requirements.txt
streamlit==0.84.0
Pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
timm==0.4.12
requests==2.26.0
python-dotenv==0.19.1
